backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database drivers
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
duckdb==0.9.2
neo4j==5.14.1
redis==5.0.1

# Vector search and embeddings
pgvector==0.2.4
sentence-transformers==2.2.2
openai==1.3.7
google-generativeai==0.3.2

# Data processing
pandas==2.1.3
numpy==1.25.2
pyarrow==14.0.1

# ETL and orchestration
prefect==2.14.11
great-expectations==0.18.5

# PDF processing and text extraction
pypdf2==3.0.1
pdfplumber==0.10.0
python-multipart==0.0.6

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1

# Caching and background tasks
celery==5.3.4
tenacity==8.2.3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and environment
python-decouple==3.8
pyyaml==6.0.1

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# ML and AI
scikit-learn==1.3.2
langchain==0.0.340
langchain-community==0.0.7
ollama==0.1.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7